+++ js/cropperjs/src/js/change.js
@@ -10,8 +10,35 @@ const ACTION_SOUTH_WEST = 'sw';
 const ACTION_NORTH_EAST = 'ne';
 const ACTION_NORTH_WEST = 'nw';
 
+function getMaxZoomRatio(pointers) {
+  const pointers2 = $.extend({}, pointers);
+  const ratios = [];
+
+  $.each(pointers, (pointer, pointerId) => {
+    delete pointers2[pointerId];
+
+    $.each(pointers2, (pointer2) => {
+      const x1 = Math.abs(pointer.startX - pointer2.startX);
+      const y1 = Math.abs(pointer.startY - pointer2.startY);
+      const x2 = Math.abs(pointer.endX - pointer2.endX);
+      const y2 = Math.abs(pointer.endY - pointer2.endY);
+      const z1 = Math.sqrt((x1 * x1) + (y1 * y1));
+      const z2 = Math.sqrt((x2 * x2) + (y2 * y2));
+      const ratio = (z2 - z1) / z1;
+
+      ratios.push(ratio);
+    });
+  });
+
+  ratios.sort((a, b) => {
+    return Math.abs(a) < Math.abs(b);
+  });
+
+  return ratios[0];
+}
+
 export default {
-  change(shiftKey, originalEvent) {
+  change(e) {
     const self = this;
     const options = self.options;
     const containerData = self.containerData;
@@ -33,7 +60,7 @@ export default {
     let offset;
 
     // Locking aspect ratio in "free mode" by holding shift key
-    if (!aspectRatio && shiftKey) {
+    if (!aspectRatio && e.shiftKey) {
       aspectRatio = width && height ? width / height : 1;
     }
 
@@ -43,24 +70,21 @@ export default {
       maxWidth = minLeft + Math.min(
         containerData.width,
         canvasData.width,
-        canvasData.left + canvasData.width
+        canvasData.left + canvasData.width,
       );
       maxHeight = minTop + Math.min(
         containerData.height,
         canvasData.height,
-        canvasData.top + canvasData.height
+        canvasData.top + canvasData.height,
       );
     }
 
+    const pointers = self.pointers;
+    const pointer = pointers[Object.keys(pointers)[0]];
     const range = {
-      x: self.endX - self.startX,
-      y: self.endY - self.startY,
+      x: pointer.endX - pointer.startX,
+      y: pointer.endY - pointer.startY,
     };
-
-    if (aspectRatio) {
-      range.X = range.y * aspectRatio;
-      range.Y = range.x / aspectRatio;
-    }
 
     switch (action) {
       // Move crop box
@@ -77,11 +101,15 @@ export default {
           break;
         }
 
+        if (right + range.x > maxWidth) {
+          range.x = maxWidth - right;
+        }
+
         width += range.x;
 
         if (aspectRatio) {
           height = width / aspectRatio;
-          top -= range.Y / 2;
+          top -= (range.x / aspectRatio) / 2;
         }
 
         if (width < 0) {
@@ -98,12 +126,16 @@ export default {
           break;
         }
 
+        if (top + range.y < minTop) {
+          range.y = minTop - top;
+        }
+
         height -= range.y;
         top += range.y;
 
         if (aspectRatio) {
           width = height * aspectRatio;
-          left += range.X / 2;
+          left += (range.y * aspectRatio) / 2;
         }
 
         if (height < 0) {
@@ -120,12 +152,16 @@ export default {
           break;
         }
 
+        if (left + range.x < minLeft) {
+          range.x = minLeft - left;
+        }
+
         width -= range.x;
         left += range.x;
 
         if (aspectRatio) {
           height = width / aspectRatio;
-          top += range.Y / 2;
+          top += (range.x / aspectRatio) / 2;
         }
 
         if (width < 0) {
@@ -142,11 +178,15 @@ export default {
           break;
         }
 
+        if (bottom + range.y > maxHeight) {
+          range.y = maxHeight - bottom;
+        }
+
         height += range.y;
 
         if (aspectRatio) {
           width = height * aspectRatio;
-          left -= range.X / 2;
+          left -= (range.y * aspectRatio) / 2;
         }
 
         if (height < 0) {
@@ -212,7 +252,7 @@ export default {
           height -= range.y;
           top += range.y;
           width = height * aspectRatio;
-          left += range.X;
+          left += range.y * aspectRatio;
         } else {
           if (range.x <= 0) {
             if (left > minLeft) {
@@ -348,19 +388,7 @@ export default {
 
       // Zoom canvas
       case 'zoom':
-        self.zoom(((x1, y1, x2, y2) => {
-          const z1 = Math.sqrt((x1 * x1) + (y1 * y1));
-          const z2 = Math.sqrt((x2 * x2) + (y2 * y2));
-
-          return (z2 - z1) / z1;
-        })(
-          Math.abs(self.startX - self.startX2),
-          Math.abs(self.startY - self.startY2),
-          Math.abs(self.endX - self.endX2),
-          Math.abs(self.endY - self.endY2)
-        ), originalEvent);
-        self.startX2 = self.endX2;
-        self.startY2 = self.endY2;
+        self.zoom(getMaxZoomRatio(pointers), e);
         renderable = false;
         break;
 
@@ -372,8 +400,8 @@ export default {
         }
 
         offset = $.getOffset(self.cropper);
-        left = self.startX - offset.left;
-        top = self.startY - offset.top;
+        left = pointer.startX - offset.left;
+        top = pointer.startY - offset.top;
         width = cropBoxData.minWidth;
         height = cropBoxData.minHeight;
 
@@ -400,7 +428,7 @@ export default {
 
         break;
 
-      // No default
+      default:
     }
 
     if (renderable) {
@@ -414,7 +442,9 @@ export default {
     }
 
     // Override
-    self.startX = self.endX;
-    self.startY = self.endY;
+    $.each(pointers, (p) => {
+      p.startX = p.endX;
+      p.startY = p.endY;
+    });
   },
 };